TEST-64 Fix code smells and hotspots
