+++ noteally_app/webservices/ws_materials.py
@@ -1,12 +1,8 @@
-# Load the rest framework libraries
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
-
-# Load the serializers
+from noteally_app.CustomPagination import CustomPagination
 from noteally_app.serializers import PostMaterialSerializer, MaterialSerializer
-
-# Load the models
 from noteally_app.models import Material
 
 
@@ -56,10 +52,13 @@ def get_materials(request):
         materials = materials.order_by(request.GET["order_by"])
     else:
         materials = materials.order_by('-total_downloads', '-total_likes')
-    
-    serializer = MaterialSerializer(materials, many=True)
-    return Response(serializer.data)
 
+    paginator = CustomPagination()
+    paginated_materials = paginator.paginate_queryset(materials, request)
+    serializer = MaterialSerializer(paginated_materials, many=True)
+    paginated_response = paginator.get_paginated_response(serializer.data)
+    
+    return Response(paginated_response.data, status=status.HTTP_200_OK)
 
 
 @api_view(['GET', 'POST'])
