+++ sonar-project.properties
@@ -10,6 +10,6 @@ sonar.projectVersion=1.0
 sonar.sourceEncoding=UTF-8
 
 # Path to coverage exclusion files
-sonar.coverage.exclusions=**/manage.py,**/NoteAlly/**,**/noteally_app/models.py,**/noteally_app/admin.py,**/noteally_app/apps.py, **/noteally_app/tests/**, **/noteally_app/views.py, **/noteally_app/serializers.py
+sonar.coverage.exclusions=**/manage.py,**/NoteAlly/**,**/noteally_app/models.py,**/noteally_app/admin.py,**/noteally_app/apps.py, **/noteally_app/tests/**, **/noteally_app/views.py, **/noteally_app/serializers.py, **/noteally_app/CustomPagination.py, **/noteally_app/webservices/ws_test.py
 
 sonar.python.coverage.reportPaths=coverage.xml
